replace fragments in full screen with layouts; attempt to reduce code duplication in form delegates
