--- app/src/main/res/layout/fragment_step_quiz_code_fullscreen_playground.xml
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/codeLayoutContainer"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:animateLayoutChanges="true"
-    android:orientation="vertical">
-
-    <org.stepic.droid.code.ui.CodeEditorLayout
-        android:id="@+id/codeStepLayout"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:paddingBottom="80dp"
-        android:fillViewport="true" />
-
-    <TextView
-        android:id="@+id/codeChosenLanguage"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentRight="true"
-        android:textColor="@color/material_grey"
-        android:gravity="end"
-        android:paddingEnd="16dp"
-        android:paddingRight="16dp"
-        android:textSize="12sp"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        tools:ignore="RtlSymmetry" />
-
-    <include
-        android:id="@+id/codeSubmitButton"
-        layout="@layout/view_step_quiz_submit_button"
-        android:layout_width="match_parent"
-        android:layout_height="@dimen/step_submit_button_height"
-
-        android:layout_margin="16dp"
-        android:layout_alignParentBottom="true"/>
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="1dp"
-        android:layout_above="@id/codeSubmitButton"
-        android:background="@color/step_quiz_separator" />
-
-
-</RelativeLayout>
