+++ app/src/main/res/layout/layout_step_quiz_code.xml
@@ -125,33 +125,6 @@
 
             tools:text="python3 dasjdlkas dadlj aldj akjsdlakj lkjasdkl jkdas" />
 
-        <android.support.v7.widget.AppCompatImageView
-            android:id="@+id/stepQuizActionFullscreen"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-
-            android:layout_marginTop="8dp"
-            android:layout_marginBottom="8dp"
-            android:layout_marginRight="48dp"
-            android:layout_marginEnd="48dp"
-            android:layout_gravity="center_vertical|end"
-            android:padding="8dp"
-            android:background="?selectableItemBackgroundBorderless"
-
-            app:srcCompat="@drawable/ic_step_quiz_fullscreen" />
-
-        <android.support.v7.widget.AppCompatImageView
-            android:id="@+id/stepQuizActionMore"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-
-            android:layout_margin="8dp"
-            android:layout_gravity="center_vertical|end"
-            android:padding="4dp"
-            android:background="?selectableItemBackgroundBorderless"
-
-            app:srcCompat="@drawable/ic_action_more_vert" />
-
         <View
             android:layout_width="match_parent"
             android:layout_height="1dp"
@@ -159,11 +132,13 @@
             android:layout_gravity="bottom" />
     </FrameLayout>
 
+
     <org.stepic.droid.code.ui.CodeEditorLayout
         android:id="@+id/codeStepLayout"
         android:layout_width="match_parent"
         android:layout_height="@dimen/embedded_code_editor_height"
         android:fillViewport="true" />
+
 
     <View
         android:layout_width="match_parent"
