+++ app/src/main/res/layout/layout_step_quiz_code_fullscreen_instruction.xml
@@ -13,8 +13,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_gravity="top"
-            android:padding="16dp"
-            android:visibility="gone" />
+            android:padding="16dp" />
 
         <FrameLayout
             android:id="@+id/stepQuizCodeDetails"
