+++ app/src/main/res/layout/layout_step_quiz_code_fullscreen_playground.xml
@@ -1,0 +1,69 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/codeLayoutContainer"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:animateLayoutChanges="true"
+    android:orientation="vertical">
+
+    <FrameLayout
+        android:id="@+id/stepQuizActions"
+        android:layout_width="match_parent"
+        android:layout_alignParentTop="true"
+        android:layout_height="48dp">
+        <android.support.v7.widget.AppCompatTextView
+            android:id="@+id/stepQuizActionChangeLang"
+            android:layout_width="wrap_content"
+            android:layout_height="48dp"
+            android:background="?selectableItemBackground"
+            android:lines="1"
+            android:ellipsize="end"
+
+            android:textSize="16sp"
+            android:textColor="@color/new_accent_color"
+            android:gravity="center_vertical"
+            android:layout_gravity="center_vertical|end"
+
+            android:paddingLeft="16dp"
+            android:paddingRight="16dp"
+
+            android:drawablePadding="12dp"
+
+            tools:drawableEnd="@drawable/ic_arrow_bottom"
+
+            tools:text="python3 dasjdlkas dadlj aldj akjsdlakj lkjasdkl jkdas" />
+
+        <View
+            android:layout_width="match_parent"
+            android:layout_height="1dp"
+            android:background="@color/step_quiz_separator"
+            android:layout_gravity="bottom" />
+    </FrameLayout>
+
+    <org.stepic.droid.code.ui.CodeEditorLayout
+        android:id="@+id/codeStepLayout"
+        android:layout_below="@id/stepQuizActions"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:paddingBottom="80dp"
+        android:fillViewport="true" />
+
+    <include
+        android:id="@+id/codeSubmitButton"
+        layout="@layout/view_step_quiz_submit_button"
+        android:layout_width="match_parent"
+        android:layout_height="@dimen/step_submit_button_height"
+
+        android:layout_margin="16dp"
+        android:layout_alignParentBottom="true"/>
+
+    <View
+        android:layout_width="match_parent"
+        android:layout_height="1dp"
+        android:layout_above="@id/codeSubmitButton"
+        android:background="@color/step_quiz_separator" />
+
+
+</RelativeLayout>
